Am realizat stergerea bazei de date

+++ application/admin/delete_db.py
@@ -4,10 +4,11 @@ import easygui
 
 
 def delete_database():
-    # TODO sa repar rutina astfel incat sa-mi stearga fisierul
-    db_path = os.path.join(os.path.dirname(__file__), 'matches.db')
+    folder = os.getcwd()
+    db_path = ''.join([folder, '\\application\\db\\matches.db'])
     db_uri = 'sqlite:///{}'.format(db_path)
-    easygui.msgbox(db_uri)
-    print(db_uri)
-    drop_database(db_uri)
-    easygui.msgbox("Gata")
+    if os.path.exists(db_path):
+        drop_database(db_uri)
+        easygui.msgbox("Gata")
+    else:
+        easygui.msgbox('Baza de date nu exista')
